+++ FreeDOM 0.4/Game.h
@@ -9,22 +9,22 @@ class Game
 private:
 	string name;
 	string about;
-	float price;
+	double price;
 	int index;
 	string devName;
 public:
-	Game(string name, string about, float price, int index, string devName);
+	Game(string name, string about, double price, int index, string devName);
 	Game();
 	//getters
 	string getName();
 	string getAbout();
-	float getPrice();
+	double getPrice();
 	int getIndex();
 	string getDevName();
 	//setters
 	void setName(string name);
 	void setAbout(string about);
-	void setPrice(float price);
+	void setPrice(double price);
 	void setDevName(string devName);
 };
